134042: removing & moving classes
